+++ app/package.json
@@ -10,7 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@moondreamsdev/dreamer-ui": "^1.7.6-test.0",
+    "@moondreamsdev/dreamer-ui": "^1.7.7-test.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
