--- app/src/toast/index.ts
@@ -1,5 +1,0 @@
-export { Toast } from './Toast';
-export type { ToastAction, ToastData, ToastType } from './Toast';
-export { ToastProvider } from './ToastProvider';
-export { useToast } from './useToast';
-export type { AddToastOptions, ToastContextValue } from './useToast';
