+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@moondreamsdev/dreamer-ui",
-  "version": "1.7.6",
+  "version": "1.7.7-test.1",
   "publishConfig": {
     "access": "public"
   },
