+++ lib/src/components/toast/index.ts
@@ -1,0 +1,2 @@
+export { Toast } from './Toast';
+export type { ToastAction, ToastData, ToastType } from './Toast';
