+++ lib/src/hooks/index.ts
@@ -1,1 +1,2 @@
 export { ActionModalContext, useActionModal } from './useActionModal';
+export { useToast, type AddToastOptions, type ToastContextValue } from './useToast';
