+++ lib/src/symbols/InfoCircled.tsx
