+++ lib/src/symbols/index.ts
@@ -1,8 +1,10 @@
 export { default as Check } from './Check';
 export { default as CheckCircled } from './CheckCircled';
 export { default as ChevronDown } from './ChevronDown';
+export { default as CrossCircled } from './CrossCircled';
 export { default as ExclamationTriangle } from './ExclamationTriangle';
 export { default as EyeClosed } from './EyeClosed';
 export { default as EyeOpened } from './EyeOpened';
-export { default as X } from './X';
+export { default as InfoCircled } from './InfoCircled';
 export { default as QuestionMarkCircled } from './QuestionMarkCircled';
+export { default as X } from './X';
